requirements.txt
Flask==2.3.3
paramiko==3.3.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
faker==19.6.2
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
psutil==5.9.5
gunicorn==21.2.0
requests==2.31.0
pyotp==2.9.0
publicsuffix2==2.20191221
boto3==1.34.0
selenium==4.16.0
webdriver-manager==4.0.1
selenium-stealth==1.0.6
fake-useragent==1.4.0
selenium-wire==5.1.0
pydo>=0.1.0
